REST/SSH API to extract project contributors
Initial implementation of the project contributors list extracting
via REST and SSH.

Differently from the standard Gerrit API, list of entries is returned
as multi-line JSON instead of a unique collection in a single line.
The rationale is on the potential size of the data and the underlying
processing system which is based on splitting, shuffling and sorting.

Change-Id: I96b846d732f81cc35afb6520b4c3130f5cb8b9bd