Update patch set 4
Patch Set 4: Verified+1 Code-Review+1

reinstated patchset2 since it is still faster than the attempt to use Gerrit caches.
If we really want to do that way we need to heavily refactor a lot of things, and the results are not guaranteed to be that elegant or fast.

Patch-set: 4
Reviewer: Gerrit User 1023193 <1023193@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=+1
Label: Code-Review=+1