Added branches array to json output.
For each output line it is computed the set of branches to which
the commits belong.
This is done holding a labeled table of all commits with branchnames
attached. The algorithm is quite fast: labeling one of the hugest git
repository i.e. linux repository with ~800M commits takes only 9 seconds
on my i7 4x2 core laptop.

Change-Id: I434497b9b6610ce68891465be5e7aac6147834b6