Update patch set 2
Patch Set 2: Verified-1

(11 comments)

I know this is a WIP :-) but see my comments.
I believe we have a huge lack of optimisation here because we navigate the commits tree multiple times.

Example: if we have 10 branches with roughly the same commits, which is mostly the case, because most of the commits have master in common, we navigate the same tree 10 times.

Patch-set: 2
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Verified=-1