Promote aggregation key to case class
Use a case class rather than a string as an aggregation key.
This avoids string parsing strings when extracting tokens
from the aggregation key.

It should simplify the addition of new aggregation strategies
since the handling of the tokens of the aggregation key will
be handled by the AggregationKey case class.

Change-Id: Ibd91665ea3d126f64aa14f270c3e8a09552b8d23