Improve performance of branch extraction
This change improves the performance of contributors endpoint by making
use of an in-memory/on-disk cache to avoid computing file diff for the same
objectIds over and over again for different branches.

The use of parallel collection has also been removed, as it has been
proven to greatly diminish the throughput by increasing overhead in
memory management.

Performance improves more for bigger repositories that have more branches.
For example:

- Extracting analytics, including branches, since 2000-01-01 from the
Gerrit Code Review repo, which at the time of writing has a size of 130Mb
and has only 15 branches might save up to 10 seconds. Arguably not that
much.

- Extracting analytics, including branches, since 2017-01-01 from the
platform/prebuilts/tools repo (part of AOSP), which at the time of
writing has a size of 22Gb and 215 branchea, might help latency to go
down from 7minutes to 20 seconds (20 times faster).

Extracting from this very same repo since 2000-01-01, allows to get a
response in a matter of minutes rather several hours.

Bug: Issue 10729
Change-Id: I991a5fc82d7c32c6e035da8b90ba4bebeab50188