Improve performance of branch extraction
This change improves the performance of contributors endpoint by making
use of an in-memory cache to avoid computing file diff for the same
objectIds over and over again for different branches.

The use of parallel collection has also been removed, as it has been
proven to greatly diminish the throughput by increasing overhead in
memory management.

The margin of improvement is proportional to the number of branches of
the repo.

Bug: Issue 10729
Change-Id: I991a5fc82d7c32c6e035da8b90ba4bebeab50188