Build analytics plugin aganst 3.0.0-rc0 plugin API
Note, that the plugin API 3.0.0-rc0 was not yet published to Maven
central. Therefore, to build this change, the plugin API has to be
published locally:

  $ tools/maven/api.sh install
  $ sbt assembly

OutputFormat class was moved to its own top level package. gwtjsonrpc
dependency was removed from gerrit core and thus from the plugsin API.
Borrow the JavaSqlTimestampHelper class from gwtjsonrpc project for
now and consider to rewrite it in Scala later. Copyright eader is
preserved as is. We can do that as the license is yhe same as in this
project.

Bug: Issue 10739
Change-Id: I1f720a924fdb1de39edc608a5d2413638535bb75