Build analytics plugin aganst 3.0.0-rc0 plugin API
Note, that the plugin API 3.0.0-rc0 was not yet published to Maven
central. Therefore, to build this change, the plugin API has to be
published locally:

  $ tools/maven/api.sh install
  $ sbt assembly

OutputFormat class was moved to its own top level package. gwtjsonrpc
dependency was removed from gerrit core and thus from the plugsin API.
JavaSqlTimestampHelper was copied in json package in gerrit core.

Bug: Issue 10739
Change-Id: I1f720a924fdb1de39edc608a5d2413638535bb75