Stop expensive conversions from scala to java lists
Java implementation of Gson does not know how to serialize scala
collections out of the box.  For this reason CommitsStatistics
and CommitInfo objects had been modeled with java.util.List and
java.util.Set instead, just to help gson to serialize them.

This change avoids unneded conversion from java to scala and vice-versa
by:
- Remodeling of CommitsStatistics and CommitInfo to use scala native
collections.
- Registering a scala iterable serializer to Gson so that it knows how
to serialize scala lists and sets.

This bumps up the latency of the contributors endpoint of around 50%.

For example, when extracting the entire history of master branch for
Gerrit repo, latency decreases from 58s to 27s.

Change-Id: I1ba7e81eb81ebcffaa431a4340fb9da09e6a5976