Move botlike-filename-regexp to analytics.config
The extraction of the analytics from the projects is a very expensive
operation. While the time-based parameters make sense to be in the API
the global options for "tagging" the commits as bot-like or not, is
very unlikely to change for every single invocation.

Move the botlike-filename-regexp into the new analytics.config so
that the results of the analytics extraction can be later on cached
consistently and reused as much as possible.

The actual caching is addressed in a follow-up change.

Change-Id: Id0455ca523a2fa73f9e37591d25b73d06180ae92