Start Gerrit daemon during tests
Create an actual Gerrit daemon for running the analytics
plugin tests, so that the overall runtime environment can
be a lot more similar to the target one.

Creating a Gerrit daemon does not allow to run tests in
parallel, otherwise it would risk to create multiple daemons
concurrently that will eventually fail.

Change-Id: If0b302b98002d6ef37b9d4934a3db3d3c3542784