Use Gerrit repositories for analytics tests
Mutate Gerrit bare repositories and use them for the execution
of the analytics plugin.
This makes the use-case scenarios a lot more similar to the real
runtime of the plugin, which runs inside a Gerrit daemon.

Remove all references to the copy&paste GitTestCase from Gitective
and replace it with a simple pimping of the current Gerrit TestRepository
class.

Change-Id: Ib6f09a5fbd73aafdb1eefc7a2b521f35df7c23e0