Move extract-issues to analytics.config
This commit fixes an issue whereby the extract-issue argument was
ignored when passed as parameter of the contributors endpoint.

extraction arguments that are unlikely to change often across invocation
are moved to the analytics.config instead, similarly to what has been
done for bot-like-commits.

The code now honours the extract-issue value, by reading it from the
config.

Bug: Issue 10951
Change-Id: I065c5592c9a9982d9d623f744a141ca12338aae6