Introduce ability to ignore binary files
The analytics plugin can now be configured to ignore binary files and do
not account for them when extracting contributors statistics.

Computing whether a file is binary is a quite expensive operation,
for this reason it is also possible to configure the plugin to cache the
result of that computation (either in memory or on disk), which could further
increase performances.

Feature: Issue 10756
Change-Id: I39f899a705bc1ce2e9338028e64c96785e9cc0e0