Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review-1
> 
> (15 comments)
> 
> Like the idea, but the implementation is way too big.
> Why don't we just filter by name? At the end of the day, this is also the way that Git LFS works and sounds good as the first step, IMHO.
> 
> WDYT?

I thought about it, but the problem with filter by name is that you need to have an intimate knowledge of the repo you are extracting analytics from, which might not necessarily be true, from case to case, and I don't think we should make that assumption for any user of this plugin.

It might be true for obvious suffixes (.png, .ko, etc), but it won't be for many unpredictable suffixes (AOSP, for example, has *many* unconventional ones: .core_dump, .kernel, .snapshot, etc).
It would be difficult to work out the full set of suffixes the need to be computed and that list would always need to be maintained as new repos get added.

Why do you would you prefer a static list to approach to the problem?

If it's just for the size of this change, I could look into separating the detection of the binary files, from their caching, into two separate changes.

If it's for something else, I'd like to understand drive behind a bit more.

WDYT?

Patch-set: 2
Reviewer: Gerrit User 1072905 <1072905@173816e5-2b9a-37c3-8a2e-48639d4f1153>