Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2: Code-Review-1
> > 
> > (15 comments)
> > 
> > Like the idea, but the implementation is way too big.
> > Why don't we just filter by name? At the end of the day, this is also the way that Git LFS works and sounds good as the first step, IMHO.
> > 
> > WDYT?
> 
> I thought about it, but the problem with filter by name is that you need to have an intimate knowledge of the repo you are extracting analytics from, which might not necessarily be true, from case to case, and I don't think we should make that assumption for any user of this plugin.

Actually, you do need intimate knowledge of the repo to understand which meaningful feature you want to extract. For analytics, there is no "one-size-fits-all" KPI I'm afraid.

> It might be true for obvious suffixes (.png, .ko, etc), but it won't be for many unpredictable suffixes (AOSP, for example, has *many* unconventional ones: .core_dump, .kernel, .snapshot, etc).
> It would be difficult to work out the full set of suffixes the need to be computed and that list would always need to be maintained as new repos get added.

The list can be indeed maintained inside the refs/meta/config of the repo itself. We could have an analytics command that performs an analysis of the BLOBs and suggests an initial set of extensions that can be ignored.

At the end of the day, if we understand that the file 'foo.dmg' is a binary file, then all the .dmg will be also binary files.

> Why do you would you prefer a static list to approach to the problem?

The static list is faster, smaller and easier to implement.
We can still make that list "automatically proposed" by the analytics plugin, but that is not this change and can be done separately.

> If it's just for the size of this change, I could look into separating the detection of the binary files, from their caching, into two separate changes.

It's not just the size of the change but also the complexity introduced and yet another cache to store and maintain in memory.

So in a nutshell what I'm proposing is the separation of two concerns:

1) Skipping binary files by their name matching a set of regexps.

2) Defining the per-project set of binary files regexps, possibly with some help of auto-detection by the analytics plugin.

That would achieve speed, even faster than the current solution. It's also going to be easier to implement and less complex to maintain.

Patch-set: 2
Reviewer: Gerrit User 1006192 <1006192@173816e5-2b9a-37c3-8a2e-48639d4f1153>