Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> > Patch Set 2:
> > 
> > > Patch Set 2: Code-Review-1
> > > 
> > > (15 comments)
> > > 
> > > Like the idea, but the implementation is way too big.
> > > Why don't we just filter by name? At the end of the day, this is also the way that Git LFS works and sounds good as the first step, IMHO.
> > > 
> > > WDYT?
> > 
> > I thought about it, but the problem with filter by name is that you need to have an intimate knowledge of the repo you are extracting analytics from, which might not necessarily be true, from case to case, and I don't think we should make that assumption for any user of this plugin.
> 
> Actually, you do need intimate knowledge of the repo to understand which meaningful feature you want to extract. For analytics, there is no "one-size-fits-all" KPI I'm afraid.

I totally agree on this, that's why I don't think *one list* of suffixes fits all repos and that's why I chose a dynamic approach to the problem.

If by knowing intimately the repo, you mean, knowing how much history it has, how many branches, how many contributors, etc, I agree but that I don't think that has anything to do with the detection of binary files, so I think it falls in a different context.

> 
> > It might be true for obvious suffixes (.png, .ko, etc), but it won't be for many unpredictable suffixes (AOSP, for example, has *many* unconventional ones: .core_dump, .kernel, .snapshot, etc).
> > It would be difficult to work out the full set of suffixes the need to be computed and that list would always need to be maintained as new repos get added.
> 
> The list can be indeed maintained inside the refs/meta/config of the repo itself. We could have an analytics command that performs an analysis of the BLOBs and suggests an initial set of extensions that can be ignored.

Yes that could be done, but maintaining a list of suffixes inside the refs/meta/config, nevertheless, is maintenance, isn't.
Even if we (write and) use a tool/command that suggests an initial set of extensions, that would still be another thing to do, every time a new project is added (or we need to write the code to automate it). So I don't think it diminishes the maintenance burden to be honest.

> 
> At the end of the day, if we understand that the file 'foo.dmg' is a binary file, then all the .dmg will be also binary files.

I like the idea, I think is a safe assumption to make.
But this could be applied even to the current approach, no?

> 
> > Why do you would you prefer a static list to approach to the problem?
> 
> The static list is faster, smaller and easier to implement.
> We can still make that list "automatically proposed" by the analytics plugin, but that is not this change and can be done separately.
> 
> > If it's just for the size of this change, I could look into separating the detection of the binary files, from their caching, into two separate changes.
> 
> It's not just the size of the change but also the complexity introduced and yet another cache to store and maintain in memory.

I understand, but wouldn't we need to maintain the list of static suffixes in a cache anyway, even if we changed approach?

> 
> So in a nutshell what I'm proposing is the separation of two concerns:
> 
> 1) Skipping binary files by their name matching a set of regexps.
> 
> 2) Defining the per-project set of binary files regexps, possibly with some help of auto-detection by the analytics plugin.
> 
> That would achieve speed, even faster than the current solution. It's also going to be easier to implement and less complex to maintain.

as I stated above, I am still not 100% convinced this is easier to implement (still needs to be done) or less complex to maintain (update list of suffixes).
Having said that, I understand where you come from and the reason you are suggesting this approach, I trust your experience on this matter, so I am happy to change this accordingly as soon as I am back on working on this.

Patch-set: 2
Reviewer: Gerrit User 1072905 <1072905@173816e5-2b9a-37c3-8a2e-48639d4f1153>