Add ability to ignore specific files suffixes
Introduce a new 'ignore-file-suffix' configuration in the
analytics.config file to ignore matching files from the statistics.

A typical case for this would be to ignore binary files and avoid
accounting for them when extracting contributor statistics.

This conveniently speeds up performances since skips completely the
loading and the diffing of binary files, which is a quite an expensive
computation.

Feature: Issue 10756
Change-Id: I60deb05586b7fd984f3067e3b3df465cd46597a7