Evict the project cache for reloading the project.config
Update Gerrit API to 3.3.0-SNAPSHOT and allow to reload
the project config by evicting the project cache explicitly.

Gerrit v3.3 introduces a new project cache and includes
tha ability to serialize its state to a persistent storage.

This allows to workaround the Issue 13754 and still being
able to add commentLink for being used in the tests.

Change-Id: I0404376a778b375d4f607a86068001f49ff391ad